move `ontology-chip` to `shared` folder
