+++ packages/hash/frontend/src/pages/[account-slug]/entities/[entity-id].page/entity-page-wrapper/entity-page-header.tsx
@@ -6,7 +6,7 @@ import { useRouter } from "next/router";
 import slugify from "slugify";
 import { TopContextBar } from "../../../../shared/top-context-bar";
 import { HashOntologyIcon } from "../../../types/entity-type/hash-ontology-icon";
-import { OntologyChip } from "../../../types/entity-type/ontology-chip";
+import { OntologyChip } from "../../../shared/ontology-chip";
 
 export const EntityPageHeader = ({ entityLabel }: { entityLabel: string }) => {
   const router = useRouter();
