+++ packages/hash/frontend/src/pages/[account-slug]/types/entity-type/property-menu.tsx
@@ -21,7 +21,10 @@ import {
   usePopupState,
 } from "material-ui-popup-state/hooks";
 import { Fragment, useId } from "react";
-import { OntologyChip, parseUriForOntologyChip } from "./ontology-chip";
+import {
+  OntologyChip,
+  parseUriForOntologyChip,
+} from "../../shared/ontology-chip";
 import { mustBeVersionedUri } from "./util";
 
 export const PropertyMenu = ({
