+++ packages/hash/frontend/src/pages/[account-slug]/types/new/entity-type.page.tsx
@@ -25,7 +25,7 @@ import { useInitTypeSystem } from "../../../../lib/use-init-type-system";
 import { getPlainLayout, NextPageWithLayout } from "../../../../shared/layout";
 import { TopContextBar } from "../../../shared/top-context-bar";
 import { HashOntologyIcon } from "../entity-type/hash-ontology-icon";
-import { OntologyChip } from "../entity-type/ontology-chip";
+import { OntologyChip } from "../../shared/ontology-chip";
 import { useRouteNamespace } from "../entity-type/use-route-namespace";
 
 const FormHelperLabel = ({
